Tests for inner product.

+++ symmetric_group/character.h
@@ -76,8 +76,8 @@ struct SymmetricGroupCharacter {
     return *this;
   }
 
-  friend int64_t InnerProduct(SymmetricGroupCharacter const &lhs,
-                              SymmetricGroupCharacter const &rhs) {
+  friend double InnerProduct(SymmetricGroupCharacter const &lhs,
+                             SymmetricGroupCharacter const &rhs) {
     if (lhs.values_.empty()) { return 0; }
     int64_t result = 0;
     for (auto const &[partition, coefficient] : lhs.values_) {
@@ -85,7 +85,8 @@ struct SymmetricGroupCharacter {
       if (iter == rhs.values_.end()) { continue; }
       result += (iter->second * coefficient * CycleTypeCount(partition));
     }
-    return result / Factorial(lhs.values_.begin()->first.whole());
+    return static_cast<double>(result) /
+           Factorial(lhs.values_.begin()->first.whole());
   }
 
   friend std::ostream &operator<<(std::ostream &os,
